+++ CHANGELOG.md
@@ -9,6 +9,10 @@ And please only add new entries to the top of this list, right below the `# Unre
 # Unreleased
 
 - Bump MSRV from `1.60` to `1.64`.
+- Web: Canvases will now capture the contextmenu if prevent_default is enabled so they don't interfere with right click pointer events.
+- Web: Canvases will now have `user-select: none` to prevent double clicks from selecting text outside the canvas.
+- Web: Only use pointer events for touch typed events.
+- Web: Use mouse events for generic mouse button events.
 
 # 0.28.3